@@ -4,17 +4,17 @@
   "dependencies": {
     "@reach/router": "1.2.1",
     "antd": "3.10.3",
-    "apollo-cache-inmemory": "1.3.5",
-    "apollo-client": "2.4.2",
+    "apollo-cache-inmemory": "1.3.8",
+    "apollo-client": "2.4.5",
     "apollo-link-error": "1.1.1",
     "apollo-link-http": "1.5.5",
     "apollo-link-ws": "1.0.9",
     "axios": "0.18.0",
     "body-scroll-lock": "2.5.10",
     "classnames": "2.2.6",
-    "date-fns": "^1.29.0",
+    "date-fns": "1.29.0",
     "eventemitter3": "3.1.0",
-    "firebase": "5.5.4",
+    "firebase": "5.5.6",
     "graphql": "14.0.2",
     "graphql-tag": "2.10.0",
     "history": "4.7.2",
@@ -23,20 +23,20 @@
     "loadable-components": "2.2.3",
     "lodash": "4.17.11",
     "lscache": "1.2.0",
-    "nanoid": "1.3.0",
+    "nanoid": "1.3.1",
     "prettyjson": "1.2.1",
-    "purched-antd-icons": "^0.2.2",
-    "react": "16.5.2",
+    "purched-antd-icons": "0.2.2",
+    "react": "16.7.0-alpha.0",
     "react-adopt": "0.6.0",
     "react-apollo": "2.2.4",
     "react-content-loader": "3.2.0",
-    "react-dom": "16.5.2",
+    "react-dom": "16.7.0-alpha.0",
     "react-facebook": "6.0.14",
     "react-helmet": "5.2.0",
     "react-loadable": "5.5.0",
     "react-props-editor": "1.0.5",
     "react-scripts": "2.0.5",
-    "semantic-ui-react": "0.82.5",
+    "semantic-ui-react": "0.83.0",
     "subscriptions-transport-ws": "0.9.15"
   },
   "scripts": {
@@ -52,7 +52,7 @@
     "cross-env": "5.2.0",
     "eslint-plugin-prettier": "3.0.0",
     "prettier": "1.14.3",
-    "webpack-bundle-analyzer": "^3.0.3"
+    "webpack-bundle-analyzer": "3.0.3"
   },
   "browserslist": [
     ">0.2%",
