@@ -1,9 +1,12 @@
-import React, { Component } from "react";
+import React, { useState, useContext, useEffect } from "react";
 import { Segment } from "semantic-ui-react";
 import { GlobalConsumer } from "./../contexts";
 import { Loader } from "semantic-ui-react";
 import { Router } from "@reach/router";
 import loadable from "loadable-components";
+import Home from "./home";
+import { Context as LegacyContext } from "./../contexts/LayoutContext";
+import { Context as ResponsiveContext } from "./../contexts/Responsive";
 
 const CreatePost = loadable(() => import(`./create-post`), {
   LoadingComponent: () => <Loader inline="centered" />
@@ -20,35 +23,67 @@ export const createPageRoute = importObj => {
 };
 
 const UserList = createPageRoute("./user-list");
-const Home = createPageRoute("./home");
+// const Home = createPageRoute("./home");
 const PostFeed = createPageRoute("./post-feed");
 const TagFeed = createPageRoute("./tag-feed");
 
-export default class EcoContent extends Component {
-  state = { visible: false };
+function SidebarWrapper() {
+  const { isMobile } = useContext(ResponsiveContext);
+  const [showSidebarScroll, setState] = useState(null);
 
-  render = () => (
+  const style2 = {
+      left: "0",
+      position: "fixed",
+      top: "50px",
+      bottom: "0",
+      width: 250,
+      zIndex: 901,
+      backgroundColor: "#f3f4f5",
+      overflowY: showSidebarScroll ? "scroll" : "hidden"
+    },
+    style3 = { paddingRight: 5, width: 240 };
+
+  if (!isMobile) {
+    style3.paddingTop = 0;
+  } else {
+    style2.overflowY = "scroll";
+  }
+
+  return (
+    <div
+      style={style2}
+      onMouseEnter={() => setState(true)}
+      onMouseLeave={() => setState(null)}
+    >
+      <Segment basic style={style3}>
+        <Sidebar />
+      </Segment>
+    </div>
+  );
+}
+
+export default function() {
+  const { showSidebar, setShowSidebar } = useContext(LegacyContext);
+  const { isMobile } = useContext(ResponsiveContext);
+  const [state] = useState({});
+  const { wrapperActive } = state;
+  useEffect(
+    () => {
+      if (!isMobile) {
+        setShowSidebar(true);
+      }
+    },
+    [isMobile]
+  );
+  return (
     <GlobalConsumer>
-      {({ responsive: { isMobile }, postView: { post, closeFn } }) => {
-        const { showSidebarScroll, wrapperActive } = this.state;
-        const { showSidebar } = this.props;
-        const style1 = {},
-          style2 = {
-            left: "0",
-            position: "fixed",
-            top: "50px",
-            bottom: "0",
-            width: 250,
-            zIndex: 901,
-            backgroundColor: "#f3f4f5",
-            overflowY: showSidebarScroll ? "scroll" : "hidden"
-          },
-          style3 = { paddingRight: 5, width: 240 };
+      {({
+        responsive: { isMobile, stretched },
+        postView: { post, closeFn }
+      }) => {
+        const style1 = {};
         if (!isMobile) {
           style1.paddingLeft = 255;
-          style3.paddingTop = 0;
-        } else {
-          style2.overflowY = "scroll";
         }
         style1.paddingTop = 55;
         style1.paddingRight = isMobile ? 0 : 5;
@@ -56,21 +91,12 @@ export default class EcoContent extends Component {
           style1.visibility = "collapse";
         }
         return (
-          <div style={style1}>
-            {showSidebar && (
-              <div
-                style={style2}
-                onMouseEnter={() => this.setState({ showSidebarScroll: true })}
-                onMouseLeave={() =>
-                  this.setState({ showSidebarScroll: undefined })
-                }
-              >
-                <Segment basic style={style3}>
-                  <Sidebar />
-                </Segment>
-              </div>
-            )}
-            <Router primary={false}>
+          <div
+            className={stretched ? "content-panel" : undefined}
+            style={style1}
+          >
+            {showSidebar && <SidebarWrapper />}
+            <Router primary={false} className="content-panel">
               <UserList path="u/*" />
               <TagFeed path="t/:tagName" />
               <PostFeed path="p/*" />
