@@ -6,41 +6,31 @@ module.exports = Object.assign(
     graphqlServer: localStorage.getItem("graphql"),
     graphqlWsServer: localStorage.getItem("graphql-ws"),
     reqCredentials: "include",
-    sections: [
-      {
-        key: "question",
-        color: "yellow",
-        displayName: "Questions"
-      },
-      {
-        key: "answer",
-        color: "blue",
-        displayName: "Answers"
-      },
-      {
-        key: "buy",
-        color: "orange",
-        displayName: "Buying"
-      },
-      {
-        key: "sell",
-        color: "green",
-        displayName: "Selling"
-      }
-    ],
+    sections: [],
     contents: [
       {
         code: "buy-item",
         name: "Buying Item",
         icon: { type: "shopping-cart", theme: "outlined" }
+      },
+      {
+        code: "sell-item",
+        name: "Selling Item",
+        icon: { type: "shop", theme: "outlined" }
+      },
+      {
+        code: "text",
+        name: "Text",
+        icon: { type: "font-size", theme: "outlined" }
       }
     ],
     posts: {
-      actionButtonSize: "tiny"
+      actionButtonSize: "medium"
     },
     staticPages: {
       privacy: "",
-      about: "https://www.sustainatrade.com"
+      about: "https://www.sustainatrade.com",
+      terms: "https://www.sustainatrade.com",
     },
     swConfigUrl:
       process.env["NODE_ENV"] === "production"
