@@ -1,22 +1,35 @@
 import gql from "graphql-tag";
 import postFragment from "./PostFragment";
 
-export const POST_LIST = gql`
-  ${postFragment}
-  query($input: PostListInput) {
-    PostList(input: $input) {
-      status
-      list {
-        __typename
+export const postListOutput = `
+pageInfo{
+  hasNextPage
+}
+edges{
+  cursor
+  node{
+    id
+    _refNo
+    __typename
+    ...PostFragment
+    ... on RemovedPost {
+      isRemoved
+      post {
         ...PostFragment
-        ... on RemovedPost {
-          isRemoved
-          _refNo
-          post {
-            ...PostFragment
-          }
-        }
       }
     }
   }
+}
 `;
+
+const key = `Post`;
+const query = gql`
+  ${postFragment}
+  query ${key} ($input: PostListInput) {
+    PostList(input: $input) {
+      ${postListOutput}
+    }
+  }
+`;
+
+export const POST_LIST = { key, query };
