@@ -1,36 +1,32 @@
 import React from "react";
 // import TagList from "./TagList";
 // import { Segment } from "semantic-ui-react";
-import { sections } from "./../../config";
+import { contents } from "./../../config";
 import SectionPosts from "./SectionPosts";
 import gql from "graphql-tag";
 import postFragment from "./../../gql-schemas/PostFragment";
 import { Query } from "react-apollo";
 import get from "lodash/get";
+import snakeCase from "lodash/snakeCase";
+import { postListOutput } from "../../gql-schemas/PostList";
 
-export const SECTIONS_QUERY = gql`
+const contentsKeys = [];
+const contentsMap = {};
+contents.forEach(c => {
+  contentsKeys.push(snakeCase(c.code));
+  contentsMap[snakeCase(c.code)] = c;
+});
+
+export const CONTENTS_QUERY = gql`
   ${postFragment}
   query {
-    ${sections.map(
-      section => `
-    ${section.key}:PostList(input: {
-      section: "${section.key}"
-      skip: 0,
+    ${contents.map(
+      content => `
+    ${snakeCase(content.code)}:PostList(input: {
+      widget: "${content.code}"
       limit: 10
     }) {
-      status
-      list {
-        id
-        _refNo
-        __typename
-        ...PostFragment
-        ... on RemovedPost {
-          isRemoved
-          post {
-            ...PostFragment
-          }
-        }
-      }
+      ${postListOutput}
     }
     `
     )}
@@ -38,18 +34,24 @@ export const SECTIONS_QUERY = gql`
 `;
 
 export default class Home extends React.Component {
+  static getDerivedStateFromError(error) {
+    console.log("error"); //TRACE
+    console.log(error); //TRACE
+  }
   render() {
     return (
       <div>
-        <Query query={SECTIONS_QUERY}>
+        <Query query={CONTENTS_QUERY}>
           {({ loading, error, data }) => {
+            console.log("data"); //TRACE
+            console.log(data); //TRACE
             return (
               <div style={{ marginTop: 3 }}>
-                {sections.map(section => (
+                {contentsKeys.map(cKey => (
                   <SectionPosts
-                    key={section.key}
-                    section={section}
-                    posts={get(data, `${section.key}.list`)}
+                    key={cKey}
+                    content={contentsMap[cKey]}
+                    posts={get(data, `${cKey}.edges`)}
                   />
                 ))}
               </div>
